chore:separate the handlers from server.js
